Moves /src/inc header files to /include
