+++ include/diagnostics.h
