+++ include/temconst.h
