+++ src/TEM.cpp
@@ -53,7 +53,7 @@
 #ifdef WITHMPI
 #include <mpi.h>
 #include "../include/RestartData.h" // for defining MPI typemap...
-//#include "inc/tbc_mpi_constants.h"
+//#include "../include/tbc_mpi_constants.h"
 #endif
 
 // For managing the floating point environment
@@ -67,7 +67,7 @@
 
 
 
-#include "inc/timeconst.h"
+#include "../include/timeconst.h"
 #include "../include/ArgHandler.h"
 #include "../include/OutputEstimate.h"
 #include "TEMLogger.h"
